requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
flask>=2.3.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
sqlalchemy>=2.0.0
pytest>=7.0.0
python-dateutil>=2.8.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
azure-identity>=1.14.0
